BP契約の上限下限時間を少数２桁まで

+++ contract/models.py
@@ -269,17 +269,19 @@ class BpContract(BaseModel):
                                           help_text=u"注文書に超過単価と不足単価の計算式を表示するか")
     allowance_base = models.IntegerField(verbose_name=u"基本給")
     allowance_base_memo = models.CharField(max_length=255, blank=True, null=True, verbose_name=u"基本給メモ")
-    allowance_time_min = models.IntegerField(default=160, verbose_name=u"時間下限", help_text=u"足りないなら欠勤となる")
-    allowance_time_max = models.IntegerField(default=200, verbose_name=u"時間上限", help_text=u"超えたら残業となる")
+    allowance_time_min = models.DecimalField(default=160, max_digits=5, decimal_places=2, verbose_name=u"時間下限",
+                                             help_text=u"足りないなら欠勤となる")
+    allowance_time_max = models.DecimalField(default=200, max_digits=5, decimal_places=2, verbose_name=u"時間上限",
+                                             help_text=u"超えたら残業となる")
     allowance_time_memo = models.CharField(max_length=255, blank=True, null=True,
                                            default=u"※基準時間：160～200/月", verbose_name=u"基準時間メモ")
     calculate_type = models.CharField(default='99', max_length=2, choices=constants.CHOICE_CALCULATE_TYPE,
                                       verbose_name=u"計算種類")
     business_days = models.IntegerField(blank=True, null=True, verbose_name=u"営業日数")
-    calculate_time_min = models.IntegerField(blank=True, null=True, default=160, verbose_name=u"計算用下限",
-                                             help_text=u"欠勤手当を算出ために使われます。")
-    calculate_time_max = models.IntegerField(blank=True, null=True, default=200, verbose_name=u"計算用上限",
-                                             help_text=u"残業手当を算出ために使われます。")
+    calculate_time_min = models.DecimalField(blank=True, null=True, default=160, max_digits=5, decimal_places=2,
+                                             verbose_name=u"計算用下限", help_text=u"欠勤手当を算出ために使われます。")
+    calculate_time_max = models.DecimalField(blank=True, null=True, default=200, max_digits=5, decimal_places=2,
+                                             verbose_name=u"計算用上限", help_text=u"残業手当を算出ために使われます。")
     allowance_overtime = models.IntegerField(default=0, verbose_name=u"残業手当")
     allowance_overtime_memo = models.CharField(max_length=255, blank=True, null=True, verbose_name=u"残業手当メモ")
     allowance_absenteeism = models.IntegerField(default=0, verbose_name=u"欠勤手当")
